Split up the JSTOR service to allow it to grow without any part getting too big
