+++ lms/services/jstor/__init__.py
@@ -1,0 +1,2 @@
+from lms.services.jstor.factory import service_factory
+from lms.services.jstor.service import JSTORService
