+++ lms/services/jstor/service.py
@@ -232,19 +232,3 @@ class JSTORService:
 
         # Strip leading/trailing whitespace and duplicate spaces
         return " ".join("".join(chunks).split())
-
-
-def factory(_context, request):
-    ai_settings = (
-        request.find_service(name="application_instance").get_current().settings
-    )
-
-    app_settings = request.registry.settings
-
-    return JSTORService(
-        api_url=app_settings.get("jstor_api_url"),
-        secret=app_settings.get("jstor_api_secret"),
-        enabled=ai_settings.get("jstor", "enabled"),
-        site_code=ai_settings.get("jstor", "site_code"),
-        http_service=request.find_service(name="http"),
-    )
